Unit3 -Implementation of Column
This is, in fact, the second version due to the loss of the first one which only contains the structure of the whole project.
PS: For this project, It`s mainly aiming to design a CMS(Content Management System), which may automatically creates some amazing websites and what you need to do is only do some settings.