A Simpe News Release System
Hi, guys, this is the prototype of a simple News Release System, basically this system has two users -admins & users who want to read some news.

For admins they have to register in order to have a account, thus it allows him to access the background system and to regulate the records of news which restored in the database(also titled NewsRealseSystem). Once some intruders attempt to use the brute-force attack to violate to the background system, we have identifying code & session as countermeasures for that situation.

And for the users, their can only browse the main page of the system, which displays various types of news that may attract most of our fans. Besides users are allowed to set some custom skins so they will not be weary of the default setting and layout.

Above is the introduction of this System, by doing this system will help you to learn some controls of ASP,NET and the principles & ideas for web designs especially for those rookies like me. Please do NOT hesitate to leave any suggest or constructive criticism. Thx, :)