Unit6 -News
In this unit, we`re still focus on news(Articles), now you are able to add a new article with fckeditor, by which you can easily set the color, style, picture etc of your content. (AddNews.aspx)

Plus  on the main page for managing news or articles(News.aspx), you are allowed to search a certain article by multiple conditions especially including time(we just used another TPPS). And DELETE or MOVE in batches, which is really convenient for news` management.