Uint8 -Final Version
Hey guys, I`m so glad to say that, this whole project for our school is completly finished, although some of the function still unavailable now, but I do hope in the future I will have some time and continue to complete this project.
Well, in this final version, we have some sorts of adminstrators, for different people, they belong to different roles, thus they have imparity rights. We divide the rights into two parts. one is the right to access pages. two, the right to delete, add, update, or move.
Lastly, for each news, they have their own HTML page which stored in the html folder. If you gonna delete a news from the system, the page,picture will be automatically deleted as well.
Thanks for my teacher who supported me so far. And thank you guys for patiently following this project :+1:  Bye for now.